python-ai-backend/app/requirements.txt
uwsgi
flask
google.generativeai
python-dotenv
pandas
numpy
requests